refactor(webapp): move DestructiveActionModal to patterns/
